download_image/requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
Django==4.1.1
djangorestframework==3.14.0
h11==0.14.0
idna==3.4
outcome==1.2.0
Pillow==9.2.0
pycparser==2.21
PySocks==1.7.1
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
selenium==4.4.3
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.2
urllib3==1.26.12
webdriver-manager==3.8.3
wsproto==1.2.0
